requirements.txt
opencv-python>=4.5.0
numpy>=1.21.0
pyrealsense2>=2.50.0
pyyaml>=5.4.0
